+++ src/nvim/api/vim.c
@@ -23,6 +23,7 @@
 #include "nvim/context.h"
 #include "nvim/decoration.h"
 #include "nvim/decoration_provider.h"
+#include "nvim/drawscreen.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/eval/typval.h"
@@ -55,7 +56,6 @@
 #include "nvim/os/process.h"
 #include "nvim/popupmenu.h"
 #include "nvim/runtime.h"
-#include "nvim/screen.h"
 #include "nvim/state.h"
 #include "nvim/types.h"
 #include "nvim/ui.h"
