+++ src/nvim/api/window.c
@@ -12,12 +12,12 @@
 #include "nvim/ascii.h"
 #include "nvim/buffer.h"
 #include "nvim/cursor.h"
+#include "nvim/drawscreen.h"
 #include "nvim/ex_docmd.h"
 #include "nvim/globals.h"
 #include "nvim/lua/executor.h"
 #include "nvim/move.h"
 #include "nvim/option.h"
-#include "nvim/screen.h"
 #include "nvim/syntax.h"
 #include "nvim/vim.h"
 #include "nvim/window.h"
