+++ src/nvim/decoration.c
@@ -4,12 +4,12 @@
 #include "nvim/api/ui.h"
 #include "nvim/buffer.h"
 #include "nvim/decoration.h"
+#include "nvim/drawscreen.h"
 #include "nvim/extmark.h"
 #include "nvim/highlight.h"
 #include "nvim/highlight_group.h"
 #include "nvim/lua/executor.h"
 #include "nvim/move.h"
-#include "nvim/screen.h"
 #include "nvim/vim.h"
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
