+++ src/nvim/memline.c
@@ -48,6 +48,7 @@
 #include "nvim/buffer.h"
 #include "nvim/change.h"
 #include "nvim/cursor.h"
+#include "nvim/drawscreen.h"
 #include "nvim/eval.h"
 #include "nvim/fileio.h"
 #include "nvim/func_attr.h"
@@ -66,7 +67,6 @@
 #include "nvim/os/process.h"
 #include "nvim/os_unix.h"
 #include "nvim/path.h"
-#include "nvim/screen.h"
 #include "nvim/sha256.h"
 #include "nvim/spell.h"
 #include "nvim/strings.h"
