+++ src/nvim/sign.c
@@ -9,6 +9,7 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/cursor.h"
+#include "nvim/drawscreen.h"
 #include "nvim/edit.h"
 #include "nvim/eval/funcs.h"
 #include "nvim/ex_docmd.h"
@@ -16,7 +17,6 @@
 #include "nvim/highlight_group.h"
 #include "nvim/move.h"
 #include "nvim/option.h"
-#include "nvim/screen.h"
 #include "nvim/sign.h"
 #include "nvim/syntax.h"
 #include "nvim/vim.h"
