+++ src/nvim/syntax.c
@@ -18,6 +18,7 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/cursor_shape.h"
+#include "nvim/drawscreen.h"
 #include "nvim/eval.h"
 #include "nvim/eval/vars.h"
 #include "nvim/ex_docmd.h"
@@ -43,7 +44,6 @@
 #include "nvim/path.h"
 #include "nvim/profile.h"
 #include "nvim/regexp.h"
-#include "nvim/screen.h"
 #include "nvim/sign.h"
 #include "nvim/strings.h"
 #include "nvim/syntax.h"
