+++ src/nvim/terminal.c
@@ -48,6 +48,7 @@
 #include "nvim/buffer.h"
 #include "nvim/change.h"
 #include "nvim/cursor.h"
+#include "nvim/drawscreen.h"
 #include "nvim/eval.h"
 #include "nvim/event/loop.h"
 #include "nvim/event/time.h"
@@ -69,7 +70,6 @@
 #include "nvim/move.h"
 #include "nvim/option.h"
 #include "nvim/os/input.h"
-#include "nvim/screen.h"
 #include "nvim/state.h"
 #include "nvim/terminal.h"
 #include "nvim/ui.h"
