+++ src/nvim/ui.c
@@ -13,6 +13,7 @@
 #include "nvim/cursor.h"
 #include "nvim/cursor_shape.h"
 #include "nvim/diff.h"
+#include "nvim/drawscreen.h"
 #include "nvim/event/loop.h"
 #include "nvim/ex_getln.h"
 #include "nvim/fold.h"
@@ -32,7 +33,6 @@
 #include "nvim/os/time.h"
 #include "nvim/os_unix.h"
 #include "nvim/popupmenu.h"
-#include "nvim/screen.h"
 #include "nvim/ui.h"
 #include "nvim/ui_compositor.h"
 #include "nvim/vim.h"
